Fix bug document not defined
