+++ src/domUpdates.js
@@ -7,10 +7,15 @@ import {
 } from './charts';
 import {
   getWeekly,
+  getUserData,
   getCurrentDate,
   getDataByDate,
   getAllTimeAverage,
+  calculateTotalTimeSlept,
 } from './model';
+import { store, getAllApiData } from './scripts';
+import { setApiData } from './apiCalls';
+
 const userInfo = document.querySelector('.data-box');
 const usersName = document.querySelector('h2');
 const userStepsEl = document.querySelector('.user-steps');
@@ -121,4 +126,50 @@ export function displayWeeklyWaterIntake(userWeeklyHydrationData) {
 
 export function toggleAddSleepModal() {
   sleepModal.classList.toggle('visible');
+}
+
+export function storeSleepData() {
+  const sleepModalForm = document.querySelector('form');
+
+  const formData = new FormData(sleepModalForm);
+  const values = [...formData.entries()];
+  const formattedValues = values.reduce((acc, value) => {
+    acc[value[0]] = value[1];
+    return acc;
+  }, {});
+
+  const totalTimeSlept = calculateTotalTimeSlept(
+    formattedValues['begin-time'],
+    formattedValues['end-time'],
+  );
+
+  const user = store.getKey('user');
+  const sleepData = getUserData(
+    'sleepData',
+    store.getKey('sleepData'),
+    user.id,
+  );
+  const nextDay = new Date(getCurrentDate(sleepData));
+  nextDay.setDate(nextDay.getDate() + 1);
+  const year = nextDay.toLocaleString('default', { year: 'numeric' });
+  const month = nextDay.toLocaleString('default', { month: '2-digit' });
+  const day = nextDay.toLocaleString('default', { day: '2-digit' });
+  const date = `${year}/${month}/${day}`;
+
+  setApiData(
+    store.getAPIKey('sleep'),
+    user.id,
+    date,
+    parseFloat(totalTimeSlept),
+    Number(formattedValues['sleep-quality']),
+  )
+    .then(() => {
+      store.getKey('todaysStepDataChart').destroy();
+      store.getKey('weeklyStepDataChart').destroy();
+      store.getKey('weeklySleepDataChart').destroy();
+      store.getKey('weeklySleepQualityChart').destroy();
+      store.getKey('weeklyHydrationDataChart').destroy();
+      getAllApiData(store.getKey('user'));
+    })
+    .catch(err => console.log(err));
 }
