+++ src/model.js
@@ -1,7 +1,3 @@
-import { setApiData } from './apiCalls';
-import { toggleAddSleepModal } from './domUpdates';
-import { getAllApiData, store } from './scripts';
-
 // Utility Functions
 
 function getRandomID(array) {
@@ -112,7 +108,7 @@ export function calculateDistanceTraveled(
 
 // Sleep Data
 
-function calculateTotalTimeSlept(...times) {
+export function calculateTotalTimeSlept(...times) {
   const [start, end] = times;
   const [hour1, min1] = start.split(':');
   const [hour2, min2] = end.split(':');
@@ -135,50 +131,4 @@ function calculateTotalTimeSlept(...times) {
   }
 
   return (diffHour + diffMin / 60).toFixed(1);
-}
-
-export function storeSleepData() {
-  const sleepModalForm = document.querySelector('form');
-
-  const formData = new FormData(sleepModalForm);
-  const values = [...formData.entries()];
-  const formattedValues = values.reduce((acc, value) => {
-    acc[value[0]] = value[1];
-    return acc;
-  }, {});
-
-  const totalTimeSlept = calculateTotalTimeSlept(
-    formattedValues['begin-time'],
-    formattedValues['end-time'],
-  );
-
-  const user = store.getKey('user');
-  const sleepData = getUserData(
-    'sleepData',
-    store.getKey('sleepData'),
-    user.id,
-  );
-  const nextDay = new Date(getCurrentDate(sleepData));
-  nextDay.setDate(nextDay.getDate() + 1);
-  const year = nextDay.toLocaleString('default', { year: 'numeric' });
-  const month = nextDay.toLocaleString('default', { month: '2-digit' });
-  const day = nextDay.toLocaleString('default', { day: '2-digit' });
-  const date = `${year}/${month}/${day}`;
-
-  setApiData(
-    store.getAPIKey('sleep'),
-    user.id,
-    date,
-    parseFloat(totalTimeSlept),
-    Number(formattedValues['sleep-quality']),
-  )
-    .then(() => {
-      store.getKey('todaysStepDataChart').destroy();
-      store.getKey('weeklyStepDataChart').destroy();
-      store.getKey('weeklySleepDataChart').destroy();
-      store.getKey('weeklySleepQualityChart').destroy();
-      store.getKey('weeklyHydrationDataChart').destroy();
-      getAllApiData(store.getKey('user'));
-    })
-    .catch(err => console.log(err));
 }
